Revision to include react-hot-loader - currently failing
