+++ app/app.jsx
@@ -1,0 +1,53 @@
+import React,  { Component } from 'react';
+import { render } from 'react-dom';
+import { Provider } from 'react-redux';
+import { Router, browserHistory } from 'react-router';
+import { syncHistoryWithStore } from 'react-router-redux';
+import createRoutes from './routes';
+import * as types from './types';
+import configureStore from './store/configureStore';
+import fetchDataForRoute from './utils/fetchDataForRoute';
+
+// Grab the state from a global injected into
+// server-generated HTML
+const initialState = window.__INITIAL_STATE__;
+
+const store = configureStore(initialState, browserHistory);
+const history = syncHistoryWithStore(browserHistory, store);
+const routes = createRoutes(store);
+
+export default class App extends Component {
+  /**
+   * Callback function handling frontend route changes.
+   */
+  render () {
+    const onUpdate = () => {
+      // Prevent duplicate fetches when first loaded.
+      // Explanation: On server-side render, we already have __INITIAL_STATE__
+      // So when the client side onUpdate kicks in, we do not need to fetch twice.
+      // We set it to null so that every subsequent client-side navigation will
+      // still trigger a fetch data.
+      // Read more: https://github.com/choonkending/react-webpack-node/pull/203#discussion_r60839356
+      if (window.__INITIAL_STATE__ !== null) {
+        window.__INITIAL_STATE__ = null;
+        return;
+      }
+
+      store.dispatch({ type: types.CREATE_REQUEST });
+      fetchDataForRoute(this.state)
+        .then((data) => {
+          return store.dispatch({ type: types.REQUEST_SUCCESS, data });
+        });
+    }
+
+    // Router converts <Route> element hierarchy to a route config:
+    // Read more https://github.com/rackt/react-router/blob/latest/docs/Glossary.md#routeconfig
+    return(
+      <Provider store={store}>
+        <Router history={history} onUpdate={onUpdate}>
+          {routes}
+        </Router>
+      </Provider>
+    );
+  }
+}
