+++ package.json
@@ -52,7 +52,6 @@
     "babel-plugin-transform-react-remove-prop-types": "^0.3.3",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
-    "babel-preset-react-hmre": "^1.1.1",
     "babel-preset-stage-0": "^6.24.1",
     "bcrypt-nodejs": "0.0.3",
     "bluebird": "^3.5.0",
@@ -88,10 +87,10 @@
     "react-addons-test-utils": "^15.5.0",
     "react-dom": "^15.5.3",
     "react-helmet": "^5.0.2",
+    "react-hot-loader": "^3.0.0-beta.7",
     "react-redux": "^5.0.3",
     "react-router": "^3.0.3",
     "react-router-redux": "^4.0.0",
-    "react-transform-hmr": "^1.0.4",
     "redux": "^3.6.0",
     "redux-logger": "^3.0.1",
     "redux-thunk": "^2.0.1",
