+++ webpack/rules/javascript.js
@@ -1,19 +1,17 @@
 const PATHS = require('../paths');
 
 module.exports = ({ production = false, browser = false } = {}) => {
-  const enableHotModuleReplacement = !production && browser;
-  const createPresets = enableHotModuleReplacement => {
-    const presets = ['es2015', 'react', 'stage-0'];
-    return enableHotModuleReplacement ? ['react-hmre', ...presets]: presets;
-  };
-  const presets = createPresets(enableHotModuleReplacement);
-
+  const presets = ['es2015', 'react', 'stage-0'];
   const plugins = production ? [
       'transform-react-remove-prop-types',
       'transform-react-constant-elements',
       'transform-react-inline-elements'
   ]: [];
 
+  if (!production && browser) {
+    plugins.push('react-hot-loader/babel');
+  }
+  
   return {
     test: /\.js$|\.jsx$/,
     loader: 'babel-loader',
